fix: move bottom-sheet types to models/interface
to fix Type error: Cannot find type definition file for 'bottom-sheet'. The file is in the program because: Entry point for implicit type library 'bottom-sheet'
