+++ src/models/interface.ts
@@ -1,3 +1,4 @@
+/* API */
 export interface APIErrorType {}
 
 export class APIError extends Error {
@@ -12,3 +13,11 @@ export interface ResponseOk {
   ok: boolean
   message: string
 }
+
+/* Bottom Sheet */
+export const STATE = {
+  Default: 'default',
+  Expanded: 'expanded',
+  Collapsed: 'collapsed',
+} as const
+export type State = (typeof STATE)[keyof typeof STATE]
