--- src/types/bottom-sheet/index.ts
@@ -1,6 +1,0 @@
-export const STATE = {
-  Default: 'default',
-  Expanded: 'expanded',
-  Collapsed: 'collapsed',
-} as const
-export type State = (typeof STATE)[keyof typeof STATE]
